% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/above_min.R
\name{above_min}
\alias{above_min}
\title{Calculate area under the curve of a CI that is above a certain treatment effect}
\usage{
above_min(
  TE,
  LCL,
  min_diff,
  only_outcome = FALSE,
  only_graphdata = FALSE,
  lower_better = TRUE,
  RR = TRUE
)
}
\arguments{
\item{TE}{treatment effect}

\item{LCL}{lower confidence interval limit}

\item{min_diff}{minimum important difference}

\item{only_outcome}{whether only the outcome should be calculated. FALSE returns an additional plot}

\item{lower_better}{whether lower values are better}

\item{RR}{whether Risk-Ratio is used}
}
\description{
Calculate area under the curve of a CI that is above a certain treatment effect
}
\examples{
above_min(1.1, 0.1, 0.6, RR = FALSE)
}
