% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AR_change_from_RR.R
\name{AR_change_from_RR}
\alias{AR_change_from_RR}
\title{Calculates Absolute Risk Reduction from RR}
\usage{
AR_change_from_RR(RR, AR_c)
}
\arguments{
\item{RR}{Risk Ratio}

\item{AR_c}{Absolute Risk from the Control Group}
}
\description{
Calculates Absolute Risk Reduction from RR
}
\examples{
AR_change_from_RR(3, 0.01)
}
