% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RR_from_AR_change.R
\name{RR_from_AR_change}
\alias{RR_from_AR_change}
\title{Calculates Risk Reduction from ARR}
\usage{
RR_from_AR_change(AR_change, AR_c)
}
\arguments{
\item{ARR}{Risk Ratio}

\item{RR}{Absolute Risk from the Control Group}
}
\description{
Calculates Risk Reduction from ARR
}
\examples{
RR_from_AR_change(-0.005, 0.01)
}
